requirements.txt
requests
fake-useragent
pydub
pytest
google-generativeai
instructor
ipython
fastapi